Move entities to pluralised directories
These were originally generated in `legislation/`, `organisation/` and
`profession/` directories. However, we want the entities to live in the
same directories as the auto-generated controllers, which by default has
a pluralised directory name.
